backend/AI Model for CV Analysis and Optimization/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
PyPDF2
pdfminer.six
spacy
